improvments to PR
