+++ onto-viewer-core/src/test/java/org/edmcouncil/spec/ontoviewer/core/ontology/data/label/provider/DefaultSettingsTest.java
@@ -1,31 +1,27 @@
 package org.edmcouncil.spec.ontoviewer.core.ontology.data.label.provider;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.XPathExpressionException;
-import org.semanticweb.owlapi.apibinding.OWLManager;
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.model.impl.ViewerCoreConfiguration;
 import org.edmcouncil.spec.ontoviewer.core.ontology.OntologyManager;
 import org.junit.jupiter.api.BeforeEach;
-import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.semanticweb.owlapi.model.IRI;
 import org.semanticweb.owlapi.model.OWLOntologyCreationException;
-import org.semanticweb.owlapi.model.OWLOntologyManager;
 import org.xml.sax.SAXException;
-import org.semanticweb.owlapi.model.OWLOntology;
 
 /**
- *
  * @author patrycja.miazek (patrycja.miazek@makolab.com)
  */
-public class DefaultSettingsTest {
+public class DefaultSettingsTest extends BasicOntologyLoader{
 
   @TempDir
   Path tempDir;
@@ -34,37 +30,13 @@ public class DefaultSettingsTest {
 
   @BeforeEach
   public void setUp() throws URISyntaxException, IOException, OWLOntologyCreationException, ParserConfigurationException, XPathExpressionException, SAXException {
-
-    OWLOntologyManager owlOntologyManager = OWLManager.createOWLOntologyManager();
-
-    OntologyManager ontologyManager = new OntologyManager();
-
-    var ontologyFileName = "LabelProviderTest.owl";
-    prepareOntologyFiles(ontologyFileName);
-
-    OWLOntology ontology = owlOntologyManager
-        .loadOntologyFromOntologyDocument(
-            tempDir
-                .resolve("LabelProviderTest.owl")
-                .toFile());
-    ontologyManager.updateOntology(ontology);
-
+    OntologyManager ontologyManager = prepareOntology(tempDir);
+    
     ViewerCoreConfiguration viewerCoreConfiguration = new ViewerCoreConfiguration();
     labelProviderTest = new LabelProvider(viewerCoreConfiguration);
     labelProviderTest.setOntologyManager(ontologyManager);
-
   }
-
-  private void prepareOntologyFiles(String ontologyFileName) throws IOException, URISyntaxException {
-    var ontologyLocationPath = getClass().getResource("/integartion_test/tests_ontology");
-    if (ontologyLocationPath == null) {
-      fail("Unable to find ontology files.");
-    }
-    var ontologyLocation = Path.of(ontologyLocationPath.toURI());
-    Files.copy(ontologyLocation.resolve(ontologyFileName),
-        tempDir.resolve(ontologyFileName));
-  }
-
+  
   @Test
   public void getLabelOrDefaultFragmentTest() {
     if (labelProviderTest == null) {
@@ -80,7 +52,6 @@ public class DefaultSettingsTest {
     for (Map.Entry<String, String> entry : expectedResult.entrySet()) {
       String result = labelProviderTest.getLabelOrDefaultFragment(IRI.create(entry.getKey()));
       assertEquals(entry.getValue(), result);
-
     }
   }
 
@@ -94,7 +65,6 @@ public class DefaultSettingsTest {
     for (Map.Entry<String, String> entry : expectedResult.entrySet()) {
       String result = labelProviderTest.getLabelOrDefaultFragment(IRI.create(entry.getKey()));
       assertEquals(entry.getValue(), result);
-
     }
   }
 }
