+++ onto-viewer-core/src/test/java/org/edmcouncil/spec/ontoviewer/core/ontology/data/label/provider/ForceLabelLangTest.java
@@ -1,14 +1,14 @@
 package org.edmcouncil.spec.ontoviewer.core.ontology.data.label.provider;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.XPathExpressionException;
-import org.semanticweb.owlapi.apibinding.OWLManager;
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.model.AppConfiguration;
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.model.ConfigItemType;
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.model.ConfigKeys;
@@ -17,20 +17,16 @@ import org.edmcouncil.spec.ontoviewer.configloader.configuration.model.impl.elem
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.model.impl.element.StringItem;
 import org.edmcouncil.spec.ontoviewer.core.ontology.OntologyManager;
 import org.junit.jupiter.api.BeforeEach;
-import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.semanticweb.owlapi.model.IRI;
 import org.semanticweb.owlapi.model.OWLOntologyCreationException;
-import org.semanticweb.owlapi.model.OWLOntologyManager;
 import org.xml.sax.SAXException;
-import org.semanticweb.owlapi.model.OWLOntology;
 
 /**
- *
  * @author patrycja.miazek (patrycja.miazek@makolab.com)
  */
-public class ForceLabelLangTest {
+public class ForceLabelLangTest extends BasicOntologyLoader {
 
   @TempDir
   Path tempDir;
@@ -41,12 +37,9 @@ public class ForceLabelLangTest {
 
   @BeforeEach
   public void setUp() throws URISyntaxException, IOException, OWLOntologyCreationException, ParserConfigurationException, XPathExpressionException, SAXException {
-
-    OWLOntologyManager owlOntologyManager = OWLManager.createOWLOntologyManager();
     ViewerCoreConfiguration viewerCoreConfiguration = new ViewerCoreConfiguration();
-
-    OntologyManager ontologyManager = new OntologyManager();
-
+    OntologyManager ontologyManager = prepareOntology(tempDir);
+    
     //English is the default language, so there is no need to test it.
     StringItem labelLang = new StringItem("pl");
     viewerCoreConfiguration.addConfigElement(ConfigKeys.LABEL_LANG, labelLang);
@@ -56,29 +49,8 @@ public class ForceLabelLangTest {
     forceLabelLang.setValue(Boolean.valueOf(true));
     viewerCoreConfiguration.addConfigElement(ConfigKeys.FORCE_LABEL_LANG, forceLabelLang);
 
-    var ontologyFileName = "LabelProviderTest.owl";
-    prepareOntologyFiles(ontologyFileName);
-
-    OWLOntology ontology = owlOntologyManager
-        .loadOntologyFromOntologyDocument(
-            tempDir
-                .resolve("LabelProviderTest.owl")
-                .toFile());
-    ontologyManager.updateOntology(ontology);
-
     labelProviderTest = new LabelProvider(viewerCoreConfiguration);
     labelProviderTest.setOntologyManager(ontologyManager);
-
-  }
-
-  private void prepareOntologyFiles(String ontologyFileName) throws IOException, URISyntaxException {
-    var ontologyLocationPath = getClass().getResource("/integartion_test/tests_ontology");
-    if (ontologyLocationPath == null) {
-      fail("Unable to find ontology files.");
-    }
-    var ontologyLocation = Path.of(ontologyLocationPath.toURI());
-    Files.copy(ontologyLocation.resolve(ontologyFileName),
-        tempDir.resolve(ontologyFileName));
   }
 
   @Test
@@ -93,8 +65,6 @@ public class ForceLabelLangTest {
     for (Map.Entry<String, String> entry : expectedResult.entrySet()) {
       String result = labelProviderTest.getLabelOrDefaultFragment(IRI.create(entry.getKey()));
       assertEquals(entry.getValue(), result);
-
     }
   }
-
 }
