--- onto-viewer-core/src/test/resources/integartion_test/config/groups_config.xml
@@ -1,111 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<config>
-  <priorityList>
-    <!--This list will be used when groups is not present-->
-    <priority>http://www.w3.org/2000/01/rdf-schema#subClassOf</priority>
-    <priority>http://www.w3.org/2000/01/rdf-schema#seeAlso</priority>
-  </priorityList>
-
-  <groups>
-    <!--The order of elements in groups will be their priority-->
-    <group>
-      <groupName>Glossary</groupName>
-      <groupItem>http://www.w3.org/2000/01/rdf-schema#label</groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#prefLabel</groupItem>
-      <groupItem>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/hasMaturityLevel
-      </groupItem>
-      <groupItem>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/synonym</groupItem>
-      <groupItem>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/abbreviation
-      </groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#definition</groupItem>
-      <groupItem>http://spec.edmcouncil.org/owlnames#mdDefinition</groupItem>
-      <groupItem>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/definitionOrigin
-      </groupItem>
-      <groupItem>http://purl.org/dc/terms/source</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/directSource</groupItem>
-      <groupItem>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/adaptedFrom
-      </groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#example</groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#scopeNote</groupItem>
-      <groupItem>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/usageNote
-      </groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#editorialNote</groupItem>
-      <groupItem>http://purl.org/dc/terms/title</groupItem>
-      <groupItem>http://purl.org/dc/terms/abstract</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/keyword</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/fileAbstract</groupItem>
-      <groupItem>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/explanatoryNote
-      </groupItem>
-    </group>
-
-    <group>
-      <groupName>This ontology contains</groupName>
-      <groupItem>@viewer.internal.clazz</groupItem>
-      <groupItem>@viewer.external.clazz</groupItem>
-      <groupItem>@viewer.internal.objectProperty</groupItem>
-      <groupItem>@viewer.external.objectProperty</groupItem>
-      <groupItem>@viewer.internal.dataProperty</groupItem>
-      <groupItem>@viewer.external.dataProperty</groupItem>
-      <groupItem>@viewer.internal.instance</groupItem>
-      <groupItem>@viewer.external.instance</groupItem>
-      <groupItem>@viewer.internal.annotationProperty</groupItem>
-      <groupItem>@viewer.external.annotationProperty</groupItem>
-    </group>
-
-    <group>
-      <groupName>Ontological characteristic</groupName>
-      <groupItem>@viewer.function.usage_classes</groupItem>
-      <groupItem>@viewer.function.direct_sub_data_property</groupItem>
-      <groupItem>@viewer.function.direct_sub_object_property</groupItem>
-      <groupItem>@viewer.function.direct_sub_annotation_property</groupItem>
-      <groupItem>@viewer.function.direct_subclasses</groupItem>
-      <groupItem>@viewer.function.instances</groupItem>
-      <groupItem>@viewer.axiom.EquivalentClasses</groupItem>
-      <groupItem>@viewer.axiom.SubClassOf</groupItem>
-      <groupItem>@viewer.function.anonymous_ancestor</groupItem>
-      <groupItem>@viewer.axiom.DisjointClasses</groupItem>
-      <groupItem>@viewer.axiom.ClassAssertion</groupItem>
-      <groupItem>@viewer.axiom.SameIndividual</groupItem>
-      <groupItem>@viewer.axiom.ObjectPropertyAssertion</groupItem>
-      <groupItem>@viewer.axiom.DataPropertyAssertion</groupItem>
-      <groupItem>@viewer.axiom.SubObjectPropertyOf</groupItem>
-      <groupItem>@viewer.axiom.SubDataPropertyOf</groupItem>
-      <groupItem>@viewer.axiom.SubAnnotationPropertyOf</groupItem>
-      <groupItem>@viewer.axiom.EquivalentObjectProperties</groupItem>
-      <groupItem>@viewer.axiom.EquivalentDataProperties</groupItem>
-      <groupItem>@viewer.axiom.ObjectPropertyDomain</groupItem>
-      <groupItem>@viewer.axiom.DataPropertyDomain</groupItem>
-      <groupItem>@viewer.axiom.ObjectPropertyRange</groupItem>
-      <groupItem>@viewer.axiom.FunctionalObjectProperty</groupItem>
-      <groupItem>@viewer.axiom.DataPropertyRange</groupItem>
-      <groupItem>@viewer.axiom.AnnotationPropertyDomain</groupItem>
-      <groupItem>@viewer.axiom.AnnotationPropertyRangeOf</groupItem>
-      <groupItem>@viewer.axiom.InverseObjectProperties</groupItem>
-      <groupItem>http://purl.org/dc/terms/hasPart</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/dependsOn</groupItem>
-    </group>
-
-    <group>
-      <groupName>Meta-information</groupName>
-      <groupItem>http://purl.org/dc/terms/creator</groupItem>
-      <groupItem>http://purl.org/dc/terms/license</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/copyright</groupItem>
-      <groupItem>http://purl.org/dc/terms/contributor</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/contributor</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/moduleAbbreviation</groupItem>
-      <groupItem>http://purl.org/dc/terms/issued</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/contentLanguage</groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#changeNote</groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#note</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/fileAbbreviation</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/filename</groupItem>
-      <groupItem>http://www.w3.org/2000/01/rdf-schema#comment</groupItem>
-      <groupItem>http://www.w3.org/2000/01/rdf-schema#isDefinedBy</groupItem>
-      <groupItem>http://www.w3.org/2000/01/rdf-schema#seeAlso</groupItem>
-      <groupItem>http://www.w3.org/2004/02/skos/core#historyNote</groupItem>
-      <groupItem>http://www.omg.org/techprocess/ab/SpecificationMetadata/relatedSpecification</groupItem>
-      <groupItem>http://purl.org/dc/terms/hasVersion</groupItem>
-    </group>
-  </groups>
-</config>
