--- onto-viewer-core/src/test/resources/integartion_test/config/ignore_to_display.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<ignoreToDisplaying>
-    <ignore>@viewer.axiom.SubObjectPropertyOf</ignore>
-    <ignore>http://www.w3.org/2000/01/rdf-schema#isDefinedBy</ignore>
-    <ignore>http://spec.edmcouncil.org/owlnames#definition</ignore>
-    <ignore>http://spec.edmcouncil.org/owlnames#label</ignore>
-    <ignore>http://spec.edmcouncil.org/owlnames#synonym</ignore>
-    <ignore>http://spec.edmcouncil.org/owlnames#example</ignore>
-    <ignore>http://spec.edmcouncil.org/owlnames#explanatoryNote</ignore>
-</ignoreToDisplaying>
