--- onto-viewer-core/src/test/resources/integartion_test/config/search_config.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<properties>
-    <searchConfig>
-        <hintFields>
-            <hintField>
-                <fieldIri>http://www.w3.org/2004/02/skos/core#definition</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </hintField>
-            <hintField>
-                <fieldIri>http://spec.edmcouncil.org/owlnames#mdDefinition</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </hintField>
-            <hintField>
-                <fieldIri>@viewer.iri.fragment</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </hintField>
-            <hintField>
-                <fieldIri>http://www.w3.org/2000/01/rdf-schema#label</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </hintField>
-            <hintField>
-                <fieldIri>http://www.w3.org/2004/02/skos/core#prefLabel</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </hintField>
-            <hintField>
-                <fieldIri>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/synonym
-                </fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </hintField>
-        </hintFields>
-        <hintThreshold>0.0</hintThreshold>
-
-        <searchFields>
-            <searchField>
-                <fieldIri>http://www.w3.org/2000/01/rdf-schema#label</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>http://www.w3.org/2004/02/skos/core#prefLabel</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>http://www.w3.org/2004/02/skos/core#definition</fieldIri>
-                <fieldBoost>0.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>http://purl.org/dc/terms/abstract</fieldIri>
-                <fieldBoost>0.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/synonym
-                </fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/abbreviation
-                </fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>http://www.omg.org/techprocess/ab/SpecificationMetadata/keyword</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>@viewer.iri.fragment</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>http://www.w3.org/2004/02/skos/core#definition</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-
-            <searchField>
-                <fieldIri>http://spec.edmcouncil.org/owlnames#mdDefinition</fieldIri>
-                <fieldBoost>1.0</fieldBoost>
-            </searchField>
-        </searchFields>
-        <searchThreshold>0.0</searchThreshold>
-
-        <!-- Maximum difference from levenstein distance, qualifies to appear in search.
-        The condition for running Levenstein's algorithm is the absence of basic results. -->
-        <searchLevensteinDistance>3.0</searchLevensteinDistance>
-
-        <searchDescriptions>
-            <searchDescription>http://www.w3.org/2004/02/skos/core#definition</searchDescription>
-            <searchDescription>http://purl.org/dc/terms/abstract</searchDescription>
-        </searchDescriptions>
-    </searchConfig>
-</properties>
-
-
